Library changes?

+++ libfops.c
@@ -5,13 +5,13 @@ int createFile(const char *filename, int flags, mode_t mode){
 }
 
 ssize_t writeToFile(int fd, const void *buf, size_t count){
-    int result = syscall(SYS_write, fd, buf, count);
+    ssize_t result = syscall(SYS_write, fd, buf, count);
     closeFile(fd);
     return result;
 }
 
 ssize_t readFromFile(int fd, void *buf, size_t count){
-    int result = syscall(SYS_read, fd, buf, count);
+    ssize_t result = syscall(SYS_read, fd, buf, count);
     closeFile(fd);
     return result;
 }
